Projects/Automation/internship-ai-agent/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.122.0
html5lib==1.1
